requirements.txt
Flask
Flask-PyMongo
cloudinary
python-dotenv
gunicorn
